about page component separated into its own file
